Updated version of honeypothttpd
