+++ servers/todo_servers.txt
@@ -1,0 +1,6 @@
+TODOS
+############################################################################
+-> Rendere compatibili i server di questa directory con le istruzioni di 
+   esecuzione del file json
+-> Creare varie tipologie di server ovviamente fake 
+-> implementare almento il servizio ftp
