+++ start.py
@@ -1,6 +1,6 @@
 import argparse
 from termcolor import colored
-from honeyhttpd.start import ServerManager, VERSION, BANNER
+from honeyhttpd.start import ServerManager
 
 
 if __name__ == '__main__':
@@ -9,10 +9,8 @@ if __name__ == '__main__':
     parser.add_argument('--config', help='Path to configuration file', required=True)
 
     args = parser.parse_args()
+    # FIXME : Verificare se il file di configurazione esiste
     config_file_path = args.config
-
-    print(colored(BANNER, 'yellow'))
-    print("Welcome to HoneyHTTPd " + VERSION + "\n")
 
     manager = ServerManager(config_file_path)
     manager.start_servers()
